LayoutTests/inspector/canvas/resolveContext-webgl2-expected.txt

Tests for the Canvas.resolveContext command for WebGL2 contexts.


== Running test suite: Canvas.resolveContextWebGL2
-- Running test case: Canvas.resolveContextWebGL2.validIdentifier
PASS: Payload should have type "object".
PASS: Payload should have className "WebGL2RenderingContext".
